// Frontend/src/app/models/enums/input-type.enum.ts
export enum InputType {
    UNKNOWN = "UNKNOWN",
    INPUT_TEXT = "INPUT_TEXT",
    INPUT_TEXT_CALLBACK = "INPUT_TEXT_CALLBACK",
    INPUT_TEXT_AUTOCOMPLETE = "INPUT_TEXT_AUTOCOMPLETE",
    INPUT_TEXT_AUTOCOMPLETE_CALLBACK = "INPUT_TEXT_AUTOCOMPLETE_CALLBACK",
    INPUT_NUMBER_INT = "INPUT_NUMBER_INT",
    INPUT_NUMBER_INT_CALLBACK = "INPUT_NUMBER_INT_CALLBACK",
    INPUT_NUMBER_FLOAT = "INPUT_NUMBER_FLOAT",
    INPUT_NUMBER_FLOAT_CALLBACK = "INPUT_NUMBER_FLOAT_CALLBACK",
    INPUT_DATE = "INPUT_DATE",
    INPUT_DATE_CALLBACK = "INPUT_DATE_CALLBACK",
    INPUT_SELECT = "INPUT_SELECT",
    INPUT_SELECT_CALLBACK = "INPUT_SELECT_CALLBACK",
    INPUT_SELECT_MULTI = "INPUT_SELECT_MULTI",
    INPUT_SELECT_MULTI_CALLBACK = "INPUT_SELECT_MULTI_CALLBACK",
    INPUT_TEXTAREA = "INPUT_TEXTAREA",
    INPUT_TEXTAREA_CALLBACK = "INPUT_TEXTAREA_CALLBACK",
    RADIO_BUTTON_GROUP = "RADIO_BUTTON_GROUP",
    RADIO_BUTTON_GROUP_CALLBACK = "RADIO_BUTTON_GROUP_CALLBACK",
    CHECKBOX = "CHECKBOX",
    CHECKBOX_CALLBACK = "CHECKBOX_CALLBACK",
    SLIDE_TOGGLE = "SLIDE_TOGGLE",
    SLIDE_TOGGLE_CALLBACK = "SLIDE_TOGGLE_CALLBACK",
    FORM_ARRAY = "FORM_ARRAY"
}